revert: remove inline schematic_symbol schema and exports fix(schematic): rely on symbol_name and existing schematic_* primitives test: split tests per rule; add symbol_name parse test
